+++ apps/mcp-server/src/mcp-server/mcp-server.ts
@@ -21,7 +21,7 @@ const routes = buildRouteMap({
 export const app = buildApplication(routes, {
   name: "mcp",
   versionInfo: {
-    currentVersion: "0.2.5",
+    currentVersion: "0.2.6",
   },
 });